pyproject.toml
[tool.poetry]
name = "hueplex"
version = "0.1.0"
description = ""
authors = ["Jeremy <jeremy.schiemann@googlemail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.2"
pydantic = "^2.6.1"
uvicorn = "^0.27.1"
python-multipart = "^0.0.9"
pyyaml = "^6.0.1"
jinja2 = "^3.1.3"
httpx = "^0.27.0"
certifi = "^2024.2.2"


[tool.poetry.group.dev.dependencies]
requests = "^2.31.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
